WIP on SemVer - refactor semver logic into separate files.
